+++ src/main/java/org/zalando/nakadi/controller/EventPublishingController.java
@@ -19,6 +19,7 @@ import org.zalando.nakadi.domain.EventPublishingStatus;
 import org.zalando.nakadi.exceptions.NakadiException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
+import org.zalando.nakadi.exceptions.runtime.EventTypeTimeoutException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.metrics.EventTypeMetricRegistry;
 import org.zalando.nakadi.metrics.EventTypeMetrics;
@@ -110,6 +111,9 @@ public class EventPublishingController {
         } catch (final NoSuchEventTypeException e) {
             LOG.debug("Event type not found.", e.getMessage());
             return create(e.asProblem(), nativeWebRequest);
+        } catch (final EventTypeTimeoutException e) {
+            LOG.debug("Failed to publish batch", e);
+            return create(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()), nativeWebRequest);
         } catch (final NakadiException e) {
             LOG.debug("Failed to publish batch", e);
             return create(e.asProblem(), nativeWebRequest);
@@ -124,9 +128,6 @@ public class EventPublishingController {
         if (eventPublishResult.getStatus() == EventPublishingStatus.SUBMITTED) {
             final long msSpent = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startingNanos);
             final String applicationName = client.getClientId();
-
-            LOG.info("[SLO] [publishing-latency] time={} size={} count={} eventTypeName={} app={}", msSpent,
-                    totalSizeBytes, eventCount, eventTypeName, applicationName);
 
             nakadiKpiPublisher.publish(kpiBatchPublishedEventType, () -> new JSONObject()
                     .put("event_type", eventTypeName)
@@ -162,7 +163,7 @@ public class EventPublishingController {
     }
 
     private ThrowableProblem createProblem(final JSONException e) {
-        return Problem.valueOf(Response.Status.BAD_REQUEST, e.getMessage());
+        return Problem.valueOf(Response.Status.BAD_REQUEST, "Error occurred when parsing event(s). " + e.getMessage());
     }
 
     private ResponseEntity response(final EventPublishResult result) {