+++ src/main/java/org/zalando/nakadi/exceptions/runtime/IllegalClientIdException.java
@@ -1,0 +1,9 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class IllegalClientIdException extends MyNakadiRuntimeException1 {
+
+    public IllegalClientIdException(final String message) {
+        super(message);
+    }
+
+}