+++ src/main/java/org/zalando/nakadi/exceptions/runtime/InvalidEventTypeException.java
@@ -1,0 +1,8 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class InvalidEventTypeException extends MyNakadiRuntimeException1 {
+
+    public InvalidEventTypeException(final String message) {
+        super(message);
+    }
+}