+++ src/main/java/org/zalando/nakadi/exceptions/runtime/TopicCreationException.java
@@ -1,0 +1,12 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class TopicCreationException extends MyNakadiRuntimeException1 {
+
+    public TopicCreationException(final String msg) {
+        super(msg);
+    }
+
+    public TopicCreationException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+}