+++ src/test/java/org/zalando/nakadi/domain/BatchItemTest.java
@@ -22,7 +22,7 @@ public class BatchItemTest {
     }
 
     @Test
-    public void testBatchItemSizeWithMultByteChar() {
+    public void testBatchItemSizeWithMultiByteChar() {
         final BatchItem item = new BatchItem("{ \"name\": \"香港\"} ",
                 BatchItem.EmptyInjectionConfiguration.build(1, false),
                 new BatchItem.InjectionConfiguration[BatchItem.Injection.values().length],