+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.7.4] - 2018-06-13
+
+### Changed
+- Switched to strict json parsing
+
+## [2.7.3] - 2018-06-11
+
+### Changed
+- Allow to use strict json parsing for event type publishing under feature toggle
+- Reduced logging by merging SLO publishing and ACCESS_LOG fields 
+
+## [2.7.2] - 2018-05-30
+
+### Changed
+- Refactored multiple exceptions
+
 ## [2.7.1] - 2018-05-28
 
 ### Changed